Rename API scripts to highlight that these are CI APIs (in case other APIs are added later)
